ecommerce-django-react-postgresql/backend/requirements.txt
asgiref==3.5.2
boto3==1.25.0
botocore==1.28.0
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
filelock==3.8.0
gunicorn==20.1.0
jmespath==1.0.1
platformdirs==2.5.2
psycopg2==2.9.3
s3transfer==0.6.0
sqlparse==0.4.3
virtualenv==20.16.5
whitenoise==6.2.0
